+++ client/src/interfaces/Config.ts
@@ -1,3 +1,5 @@
+import { WeatherData } from '../types';
+
 export interface Config {
   WEATHER_API_KEY: string;
   lat: number;
@@ -26,4 +28,6 @@ export interface Config {
   monthSchema: string;
   showTime: boolean;
   defaultTheme: string;
+  isKilometer: boolean;
+  weatherData: WeatherData;
 }