+++ client/src/types/index.ts
@@ -1,1 +1,2 @@
 export * from './ConfigFormData';
+export * from './WeatherData';